rejigging code to use reusable default projects
