+++ tests/testscenarios.py
@@ -1,11 +1,5 @@
 """
-Test script to see if Optima works.
-
-To use: comment out lines in the definition of 'tests' to not run those tests.
-
-NOTE: for best results, run in interactive mode, e.g.
-
-python -i tests.py
+Test scenarios
 
 Version: 2016jan27
 """
@@ -45,59 +39,18 @@ blank()
 T = tic()
 
 
-## GUI test
+## Standard scenario test
 if 'standardscen' in tests:
     t = tic()
 
     print('Running standard scenarios test...')
-    from optima import Project, Parscen, Budgetscen, Coveragescen
-    from optima.defaults import defaultprogset
+    from optima import Parscen, Budgetscen, Coveragescen
+    from makedefaults import defaultproject
     from numpy import array
     
-    # Make project and store results from default sim
-    P = Project(spreadsheet='concentrated.xlsx')
-    results = P.runsim('default')
-
-    caspships = P.parsets['default'].pars[0]['condcas'].y.keys()
-
-    # Get a default progset 
-    R = defaultprogset(P, addpars=True, addcostcov=True, filterprograms=['Condoms', 'FSW programs', 'HTC', 'ART', 'OST'])
+    P = defaultproject('concentrated')
     
-    # Modify target pars and pops
-    R.programs['HTC'].rmtargetpar({'param': 'hivtest', 'pop': 'M 0-14'})
-    R.programs['HTC'].rmtargetpar({'param': 'hivtest', 'pop': 'F 0-14'})
-    R.programs['HTC'].targetpops.pop(R.programs['HTC'].targetpops.index('M 0-14'))
-    R.programs['HTC'].targetpops.pop(R.programs['HTC'].targetpops.index('F 0-14'))
-    R.updateprogset()
-
-    # Add program effects
-    R.covout['condcas'][('Clients', 'FSW')].addccopar({'intercept': (0.3,0.35), 't': 2016.0, 'Condoms':(0.45,0.55), 'FSW programs':(0.55,0.65)})
-    R.covout['condcas'][('FSW', 'Clients')].addccopar({'intercept': (0.3,0.35), 't': 2016.0, 'Condoms':(0.45,0.55), 'FSW programs':(0.55,0.65)})
-    R.covout['condcas'][('Clients', 'F 15+')].addccopar({'intercept': (0.2,0.3), 't': 2016.0, 'Condoms':(0.35,0.45)})
-    R.covout['condcas'][('F 15+','Clients')].addccopar({'intercept': (0.2,0.3), 't': 2016.0, 'Condoms':(0.35,0.45)})
-    R.covout['condcas'][('MSM', 'MSM')].addccopar({'intercept': (0.5,0.55), 't': 2016.0, 'Condoms':(0.55,0.65), 'MSM programs':(0.75,0.85)})
-    R.covout['condcas'][('M 15+', 'FSW')].addccopar({'intercept': (0.3,0.35), 't': 2016.0, 'Condoms':(0.45,0.55), 'FSW programs':(0.55,0.65)})
-    R.covout['condcas'][('FSW', 'M 15+')].addccopar({'intercept': (0.3,0.35), 't': 2016.0, 'Condoms':(0.45,0.55), 'FSW programs':(0.55,0.65)})
-    R.covout['condcas'][('M 15+', 'F 15+')].addccopar({'intercept': (0.2,0.3), 't': 2016.0, 'Condoms':(0.35,0.45)})
-    R.covout['condcas'][('F 15+', 'M 15+')].addccopar({'intercept': (0.2,0.3), 't': 2016.0, 'Condoms':(0.35,0.45)})
-    R.covout['condcas'][('F 15+', 'PWID')].addccopar({'intercept': (0.1,0.2), 't': 2016.0, 'Condoms':(0.35,0.45)})
-    R.covout['condcas'][('PWID', 'F 15+')].addccopar({'intercept': (0.1,0.2), 't': 2016.0, 'Condoms':(0.35,0.45)})
-
-    R.covout['condcom'][('Clients', 'FSW')].addccopar({'intercept': (0.6,0.65), 't': 2016.0, 'FSW programs':(0.9,0.95)})
-    R.covout['condcom'][('FSW', 'Clients')].addccopar({'intercept': (0.6,0.65), 't': 2016.0, 'FSW programs':(0.9,0.95)})
-
-    R.covout['hivtest']['FSW'].addccopar({'intercept': (0.35,0.45), 't': 2016.0, 'HTC': (0.95,0.99), 'FSW programs':(0.95,0.99)})
-    R.covout['hivtest']['MSM'].addccopar({'intercept': (0.05,0.1), 't': 2016.0, 'HTC': (0.95,0.99), 'MSM programs':(0.95,0.99)})
-    R.covout['hivtest']['Clients'].addccopar({'intercept': (0.35,0.45), 't': 2016.0, 'HTC': (0.95,0.99)})
-    R.covout['hivtest']['M 15+'].addccopar({'intercept': (0.15,0.2), 't': 2016.0, 'HTC': (0.95,0.99)})
-    R.covout['hivtest']['F 15+'].addccopar({'intercept': (0.15,0.2), 't': 2016.0, 'HTC': (0.95,0.99)})
-    R.covout['hivtest']['PWID'].addccopar({'intercept': (0.05,0.1), 't': 2016.0, 'HTC': (0.95,0.99)})
-
-    R.covout['numtx']['tot'].addccopar({'intercept': (100.0,150.0), 't': 2016.0})
-    R.covout['numost']['tot'].addccopar({'intercept': (100.0,150.0), 't': 2016.0})
-    
-    # Store this program set in the project
-    P.addprogset(R)
+    caspships = P.parsets['default'].pars[0]['condcas'].keys
     
     ## Define scenarios
     scenlist = [
@@ -234,39 +187,15 @@ if 'maxbudget' in tests:
     t = tic()
 
     print('Running maximum budget scenario test...')
-    from optima import Project, Parscen, Budgetscen, odict
-    from optima.defaults import defaultprogset
+    from optima import Budgetscen, odict
     
     ## Set up default project
-    P = Project(spreadsheet='test7pops.xlsx')
-    pops = P.data['pops']['short']
-    caspships = P.data['pships']['cas']
-    
-    R = defaultprogset(P, addpars=True, addcostcov=True, filterprograms=['Condoms', 'FSW_programs', 'HTC', 'ART'])
-    R.programs['HTC'].rmtargetpar({'param': 'hivtest', 'pop': 'M 0-14'})
-    R.programs['HTC'].rmtargetpar({'param': 'hivtest', 'pop': 'F 0-14'})
-    R.programs['HTC'].targetpops.pop(R.programs['HTC'].targetpops.index('M 0-14'))
-    R.programs['HTC'].targetpops.pop(R.programs['HTC'].targetpops.index('F 0-14'))
-    R.updateprogset()
-    R.covout['condcas'][('Clients', 'FSW')].addccopar({'intercept': (0.3,0.35), 't': 2016.0, 'Condoms':(0.45,0.55), 'FSW_programs':(0.55,0.65)})
-    R.covout['condcas'][('Clients', 'F 15+')].addccopar({'intercept': (0.2,0.3), 't': 2016.0, 'Condoms':(0.35,0.45)})
-    R.covout['condcas'][('MSM', 'MSM')].addccopar({'intercept': (0.5,0.55), 't': 2016.0, 'Condoms':(0.55,0.65), 'MSM_programs':(0.75,0.85)})
-    R.covout['condcas'][('M 15+', 'FSW')].addccopar({'intercept': (0.3,0.35), 't': 2016.0, 'Condoms':(0.45,0.55), 'FSW_programs':(0.55,0.65)})
-    R.covout['condcas'][('M 15+', 'F 15+')].addccopar({'intercept': (0.2,0.3), 't': 2016.0, 'Condoms':(0.35,0.45)})
-    R.covout['condcom'][('Clients', 'FSW')].addccopar({'intercept': (0.6,0.65), 't': 2016.0, 'FSW_programs':(0.9,0.95)})
-    R.covout['hivtest']['FSW'].addccopar({'intercept': (0.35,0.45), 't': 2016.0, 'HTC': (0.95,0.99), 'FSW_programs':(0.95,0.99)})
-    R.covout['hivtest']['MSM'].addccopar({'intercept': (0.05,0.1), 't': 2016.0, 'HTC': (0.95,0.99), 'MSM_programs':(0.95,0.99)})
-    R.covout['hivtest']['Clients'].addccopar({'intercept': (0.35,0.45), 't': 2016.0, 'HTC': (0.95,0.99)})
-    R.covout['hivtest']['M 15+'].addccopar({'intercept': (0.15,0.2), 't': 2016.0, 'HTC': (0.95,0.99)})
-    R.covout['hivtest']['F 15+'].addccopar({'intercept': (0.15,0.2), 't': 2016.0, 'HTC': (0.95,0.99)})
-    R.covout['numtx']['tot'].addccopar({'intercept': (100.0,150.0), 't': 2016.0})
-    
-    P.addprogset(name='default', progset=R)
+    P = defaultproject('generalized')
     
     ## Define scenarios
     scenlist = [
         Budgetscen(name='Current conditions', parsetname='default', progsetname='default', t=[2016], budget=P.progsets['default'].getdefaultbudget()),
-        Budgetscen(name='Unlimited spending', parsetname='default', progsetname='default', t=[2016], budget=odict([(key, 1e9) for key in R.programs.keys()])),
+        Budgetscen(name='Unlimited spending', parsetname='default', progsetname='default', t=[2016], budget=odict([(key, 1e9) for key in P.progsets['default'].programs.keys()])),
         ]
     
     # Run the scenarios
